Android XML Purser
XML stands for Extensible Mark-up Language.XML is a very popular format
and commonly used for sharing data on the internet. This code how to
parse the XML file and extract necessary information from it.

Android provides three types of XML parsers which are DOM,SAX and
XMLPullParser. Among all of them android recommend XMLPullParser because
it is efficient and easy to use. So we are going to use XMLPullParser
for parsing XML.